feat: 테스트 오류 정적호스팅 설정 및 cicd 보완
issue: #3 #6

+++ deploy/deploy_client.sh
@@ -1,13 +1,18 @@
 #!/bin/sh
 
+if [ $# -ne 1 ]; then
+    echo 'No argument'
+    exit -1
+fi
+
 BUILD_PATH="docker/data/client"
 
-aws s3 cp s3://keeper-homepage/prod/front/$(date +%Y%m%d).tar.gz $BUILD_PATH
+aws s3 cp s3://keeper-homepage/prod/front/$1.tar.gz $BUILD_PATH
 
 if [ -d $BUILD_PATH/build ]
 then
     rm -rf $BUILD_PATH/build/*
 fi
 
-tar -zxvf $BUILD_PATH/$(date +%Y%m%d).tar.gz -C $BUILD_PATH/build
-rm $BUILD_PATH/$(date +%Y%m%d).tar.gz
+tar -zxvf $BUILD_PATH/$1.tar.gz -C $BUILD_PATH/build
+rm $BUILD_PATH/$1.tar.gz
