+++ deploy/deploy_server.sh
@@ -1,18 +1,21 @@
 #!/bin/sh
 
+if [ $# -ne 1 ]; then
+    echo 'No argument'
+    exit -1
+fi
+
 BUILD_PATH="docker/data/server"
 
-aws s3 cp s3://keeper-homepage/prod/back/$(date +%Y%m%d).tar.gz BUILD_PATH$
-
-
+aws s3 cp s3://keeper-homepage/prod/back/$1.tar.gz BUILD_PATH$
 
 if [ -d $BUILD_PATH/application ]
 then
     rm -r $BUILD_PATH/dependencies $BUILD_PATH/snapshot-dependencies $BUILD_PATH/spring-boot-loader $BUILD_PATH/application
 fi
 
-tar -zxvf $BUILD_PATH/$(date +%Y%m%d).tar.gz -C server
-rm $BUILD_PATH/$(date +%Y%m%d).tar.gz
+tar -zxvf $BUILD_PATH/$1.tar.gz -C server
+rm $BUILD_PATH/$1.tar.gz
 
 docker build -t keeper-homepage-app:prod -f app.layer.dockerfile .
